feat: Add configuration for Next.js export
This commit adds configuration options for Next.js export in the `next.config.mjs` file. The `reactStrictMode` is set to true, the output is set to 'export', and the images are set to be unoptimized. Additionally, the `assetPrefix` and `basePath` are conditionally set based on the environment.

The `package.json` file is also updated to include an `export` script for Next.js export.

A new workflow file `pages.yml` is added under the `.github/workflows` directory. This workflow is responsible for deploying the Next.js application to GitHub Pages.

Signed-off-by: Bhaumik Singh <Bhaumiksingh2000@gmail.com>